refactor: BigInteger 를 이용하여 long 자료형 표현 범위를 초과하는 숫자도 처리하도록 변경

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -3,6 +3,7 @@ package calculator.controller;
 import calculator.service.CalculatorService;
 import calculator.view.InputView;
 import calculator.view.OutputView;
+import java.math.BigInteger;
 
 public class CalculatorController {
 
@@ -26,7 +27,7 @@ public class CalculatorController {
         String userInput = inputView.getUserInput();
 
         // 덧셈 결과 받기
-        int result = calculatorService.getResult(userInput);
+        BigInteger result = calculatorService.getResult(userInput);
 
         // 덧셈 결과 출력
         outputView.printAddResult(result);
